requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-filter>=24.3,<25.0
django-cors-headers>=4.4,<4.5
psycopg[binary]>=3.2,<3.3
python-decouple>=3.8,<3.9
Pillow>=10.4,<10.5
gunicorn>=22.0,<23.0
